refactor: Project 도메인 Service 인터페이스 제거 및 정렬 로직 이동
- ProjectService 인터페이스 + ProjectServiceImpl 구현체를  단일 ProjectService 구체 클래스로 통합
- 컨트롤러의 ProjectComparator 정렬 로직을  ProjectService.paginateProjects()로 이동
- service/ 하위 패키지 제거하여 다른 Light 도메인과 구조 통일
