+++ src/main/java/sopt/org/homepage/project/ProjectController.java
@@ -2,10 +2,8 @@ package sopt.org.homepage.project;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import java.util.ArrayList;
 import lombok.RequiredArgsConstructor;
 import org.springdoc.core.annotations.ParameterObject;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -16,8 +14,6 @@ import sopt.org.homepage.global.common.dto.PaginateResponseDto;
 import sopt.org.homepage.project.dto.request.GetProjectsRequestDto;
 import sopt.org.homepage.project.dto.response.ProjectDetailResponseDto;
 import sopt.org.homepage.project.dto.response.ProjectsResponseDto;
-import sopt.org.homepage.project.service.ProjectService;
-import sopt.org.homepage.project.util.ProjectComparator;
 
 @RestController
 @RequiredArgsConstructor
@@ -31,18 +27,7 @@ public class ProjectController {
     public ResponseEntity<PaginateResponseDto<ProjectsResponseDto>> getProjects(
             @ParameterObject @ModelAttribute GetProjectsRequestDto getProjectsRequestDto
     ) {
-        PaginateResponseDto<ProjectsResponseDto> projects = projectService.paginateProjects(getProjectsRequestDto);
-
-        var sortedData = new ArrayList<>(projects.getData());
-        sortedData.sort(ProjectComparator::compare);
-
-        var sortedProjects = new PaginateResponseDto<>(
-                sortedData,
-                projects.getTotalCount(),
-                projects.getLimit(),
-                projects.getCurrentPage()
-        );
-        return ResponseEntity.status(HttpStatus.OK).body(sortedProjects);
+        return ResponseEntity.ok(projectService.paginateProjects(getProjectsRequestDto));
     }
 
     @GetMapping("/{projectId}")
@@ -52,7 +37,4 @@ public class ProjectController {
     ) {
         return ResponseEntity.ok(projectService.findOne(projectId));
     }
-
 }
-
-
