+++ src/main/java/sopt/org/homepage/project/ProjectService.java
@@ -1,4 +1,4 @@
-package sopt.org.homepage.project.service;
+package sopt.org.homepage.project;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -10,16 +10,18 @@ import sopt.org.homepage.infrastructure.external.playground.PlaygroundService;
 import sopt.org.homepage.project.dto.request.GetProjectsRequestDto;
 import sopt.org.homepage.project.dto.response.ProjectDetailResponseDto;
 import sopt.org.homepage.project.dto.response.ProjectsResponseDto;
+import sopt.org.homepage.project.util.ProjectComparator;
 
 @Service
 @RequiredArgsConstructor
-public class ProjectServiceImpl implements ProjectService {
+public class ProjectService {
     private final PlaygroundService playgroundService;
     private final ArrayUtil arrayUtil;
 
-    @Override
     public PaginateResponseDto<ProjectsResponseDto> paginateProjects(GetProjectsRequestDto dto) {
         var allProjects = findAll(dto);
+        allProjects.sort(ProjectComparator::compare);
+
         var paginatedProject = arrayUtil.paginateArray(allProjects, dto.getPageNo(), dto.getLimit());
 
         return new PaginateResponseDto<>(
@@ -30,17 +32,14 @@ public class ProjectServiceImpl implements ProjectService {
         );
     }
 
-    @Override
     public List<ProjectsResponseDto> findAll(GetProjectsRequestDto dto) {
         return playgroundService.getAllProjects(dto);
     }
 
-    @Override
     public ProjectDetailResponseDto findOne(Long projectId) {
         return playgroundService.getProjectDetail(projectId);
     }
 
-    @Override
     public List<ProjectsResponseDto> findByGeneration(Integer generation) {
         var allProjects = findAll(new GetProjectsRequestDto(1, Integer.MAX_VALUE, null, null));
         return allProjects.stream()
